# packages/shared/openapi.yaml
openapi: 3.0.3
info:
  title: Gordon Ulen CPA Tax Helper API
  description: Tax preparation workflow automation system
  version: 1.0.0
  contact:
    name: RaveJedi Labs
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://api.taxhelper.com
    description: Production
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase Auth JWT token
  schemas:
...
    Address:
      type: object
      properties:
        type:
          type: string
          enum: [home, mailing, business]
        street:
          type: string
        city:
          type: string
        state:
          type: string
        zip:
          type: string

    # Banking Information
    BankingInfo:
      type: object
      properties:
        routingNumber:
          type: string
        accountNumber:
          type: string
        lastVerified:
          type: string
          format: date-time

    # Customer
    Customer:
      type: object
      required:
        - firstName
        - lastName
        - email
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        addresses:
          type: array
          items:
            $ref: '#/components/schemas/Address'
        entityType:
          type: string
          enum: [individual, s-corp, partnership, c-corp, llc, schedule-c]
        ein:
          type: string
        ssnEncrypted:
          type: string
        assignedPreparer:
          type: string
        bankingInfo:
          $ref: '#/components/schemas/BankingInfo'
        portalAccess:
          type: boolean
          default: false
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CustomerCreate:
      type: object
      required:
        - firstName
        - lastName
        - email
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        addresses:
          type: array
          items:
            $ref: '#/components/schemas/Address'
        entityType:
          type: string
          enum: [individual, s-corp, partnership, c-corp, llc, schedule-c]
        ein:
          type: string
        assignedPreparer:
          type: string
        portalAccess:
          type: boolean
          default: false
    CustomerUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        addresses:
          type: array
          items:
            $ref: '#/components/schemas/Address'
        entityType:
          type: string
          enum: [individual, s-corp, partnership, c-corp, llc, schedule-c]
        ein:
          type: string
        assignedPreparer:
          type: string
        bankingInfo:
          $ref: '#/components/schemas/BankingInfo'
        portalAccess:
          type: boolean

    # Document
...
    AppointmentType:
      type: string
      enum: [tax_prep, drop_off, pick_up, signing, consultation]
    AppointmentStatus:
      type: string
      enum: [scheduled, confirmed, completed, cancelled, no_show]
    Appointment:
      type: object
      properties:
        id:
          type: string
        customerId:
          type: string
        type:
          $ref: '#/components/schemas/AppointmentType'
        scheduledAt:
          type: string
          format: date-time
        duration:
          type: integer
          description: Duration in minutes
        assignedTo:
          type: string
        status:
          $ref: '#/components/schemas/AppointmentStatus'
        reminderSent:
          type: boolean
          default: false
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AppointmentCreate:
      type: object
      required:
        - customerId
        - type
        - scheduledAt
      properties:
        customerId:
          type: string
        type:
          $ref: '#/components/schemas/AppointmentType'
        scheduledAt:
          type: string
          format: date-time
        duration:
          type: integer
          default: 60
        assignedTo:
          type: string
        notes:
          type: string

    # Communication
    CommunicationType:
      type: string
      enum: [email, sms, call]
    CommunicationDirection:
      type: string
      enum: [inbound, outbound]
    CommunicationStatus:
      type: string
      enum: [sent, delivered, failed]
    CommunicationTrigger:
      type: string
      enum: [manual, automation, agent]
    Communication:
      type: object
      properties:
        id:
          type: string
        customerId:
          type: string
        type:
          $ref: '#/components/schemas/CommunicationType'
        direction:
          $ref: '#/components/schemas/CommunicationDirection'
        subject:
          type: string
        content:
          type: string
        sentAt:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/CommunicationStatus'
        triggeredBy:
          $ref: '#/components/schemas/CommunicationTrigger'
    CommunicationSend:
      type: object
      required:
        - customerId
        - type
        - content
      properties:
        customerId:
          type: string
        type:
          $ref: '#/components/schemas/CommunicationType'
        subject:
          type: string
        content:
          type: string
        triggeredBy:
          $ref: '#/components/schemas/CommunicationTrigger'
          default: manual

    # Kanban
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true

    # Pagination
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        hasMore:
          type: boolean

    # List responses
    CustomerList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Customer'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    AppointmentList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Appointment'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
security:
  - BearerAuth: []
paths:
...
  /api/customers:
    get:
      summary: List all customers
      operationId: listCustomers
      tags:
        - Customers
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: entityType
          in: query
          schema:
            $ref: '#/components/schemas/Customer/properties/entityType'
      responses:
        '200':
          description: List of customers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerList'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Create a new customer
      operationId: createCustomer
      tags:
        - Customers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerCreate'
      responses:
        '201':
          description: Customer created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/customers/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string

    get:
      summary: Get a customer by ID
      operationId: getCustomer
      tags:
        - Customers
      responses:
        '200':
          description: Customer details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      summary: Update a customer
      operationId: updateCustomer
      tags:
        - Customers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerUpdate'
      responses:
        '200':
          description: Customer updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete a customer
      operationId: deleteCustomer
      tags:
        - Customers
      responses:
        '204':
          description: Customer deleted
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Documents
...
  /api/appointments:
    get:
      summary: List appointments
      operationId: listAppointments
      tags:
        - Appointments
      parameters:
        - name: customerId
          in: query
          schema:
            type: string
        - name: assignedTo
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/AppointmentStatus'
        - name: startDate
          in: query
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of appointments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentList'

    post:
      summary: Create an appointment
      operationId: createAppointment
      tags:
        - Appointments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentCreate'
      responses:
        '201':
          description: Appointment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
  /api/appointments/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string

    get:
      summary: Get an appointment by ID
      operationId: getAppointment
      tags:
        - Appointments
      responses:
        '200':
          description: Appointment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
        '404':
          description: Appointment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      summary: Update an appointment
      operationId: updateAppointment
      tags:
        - Appointments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                scheduledAt:
                  type: string
                  format: date-time
                duration:
                  type: integer
                assignedTo:
                  type: string
                status:
                  $ref: '#/components/schemas/AppointmentStatus'
                notes:
                  type: string
      responses:
        '200':
          description: Appointment updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'

    delete:
      summary: Cancel an appointment
      operationId: deleteAppointment
      tags:
        - Appointments
      responses:
        '204':
          description: Appointment cancelled
        '404':
          description: Appointment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Communications
  /api/communications:
    get:
      summary: List communications
      operationId: listCommunications
      tags:
        - Communications
      parameters:
        - name: customerId
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/CommunicationType'
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of communications
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Communication'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
  /api/communications/send:
    post:
      summary: Send a communication
      operationId: sendCommunication
      tags:
        - Communications
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommunicationSend'
      responses:
        '201':
          description: Communication sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Communication'
        '400':
          description: Failed to send
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Kanban
...